requirements.txt
colorama
discord
asyncio
datetime
Augmentor
discord.py
numpy
aiohttp
nudepy
json
traceback
sys
aiohttp
time
re
numpy
random
string
os
shutil
git+https://github.com/dimitrismistriotis/profanity-check.git#egg=profanity-check